+++ src/main/kotlin/com/eatsfinder/global/oauth/client/google/dto/GoogleUserInfoResponse.kt
@@ -1,6 +1,6 @@
 package com.eatsfinder.global.oauth.client.google.dto
 
-import com.eatsfinder.domain.user.dto.OAuth2UserInfo
+import com.eatsfinder.domain.user.dto.oauth.OAuth2UserInfo
 import com.eatsfinder.domain.user.model.SocialType
 import com.fasterxml.jackson.annotation.JsonProperty
 import com.fasterxml.jackson.databind.PropertyNamingStrategies
